+++ addon/data-import.js
@@ -1,5 +1,5 @@
 /* global React ReactDOM */
-import {sfConn, apiVersion} from "./inspector.js";
+import {sfConn, apiVersion, setupColorListeners} from "./inspector.js";
 /* global initButton */
 import {csvParse} from "./csv-parse.js";
 import {DescribeInfo, copyToClipboard, initScrollTable} from "./data-load.js";
@@ -916,6 +916,7 @@ let h = React.createElement;
 class App extends React.Component {
   constructor(props) {
     super(props);
+    setupColorListeners();
     this.onApiTypeChange = this.onApiTypeChange.bind(this);
     this.onImportActionChange = this.onImportActionChange.bind(this);
     this.onImportTypeChange = this.onImportTypeChange.bind(this);
@@ -936,6 +937,8 @@ class App extends React.Component {
     this.onSkipAllUnknownFieldsClick = this.onSkipAllUnknownFieldsClick.bind(this);
     this.onConfirmPopupYesClick = this.onConfirmPopupYesClick.bind(this);
     this.onConfirmPopupNoClick = this.onConfirmPopupNoClick.bind(this);
+    this.clickDecrease = this.clickDecrease.bind(this);
+    this.clickIncrease = this.clickIncrease.bind(this);
     this.unloadListener = null;
     this.state = {templateValueIndex: -1};
   }
@@ -1115,6 +1118,25 @@ class App extends React.Component {
       removeEventListener("beforeunload", this.unloadListener);
     }
   }
+
+  changeValue(e, shouldIncrease = false) {
+    const inputTarget = document.getElementById(e.target.dataset.targetid);
+    const oldValue = +inputTarget.value;
+    if((""+oldValue) == "NaN")
+      return;
+    inputTarget.value = shouldIncrease ? oldValue + 1 : oldValue - 1;
+    // trigger the onChange listener
+    const event = new Event('input', { bubbles: true });
+    inputTarget.dispatchEvent(event);
+  }
+
+  clickDecrease(e) {
+    this.changeValue(e, false);
+  }
+  clickIncrease(e) {
+    this.changeValue(e, true);
+  }
+
   render() {
     let {model} = this.props;
     return h("div", {},
@@ -1173,7 +1195,18 @@ class App extends React.Component {
                 )
               ),
               h("a", {className: "button field-info", href: model.showDescribeUrl(), target: "_blank", title: "Show field info for the selected object"},
-                h("div", {className: "button-icon"}),
+                h("div", {className: "button-icon"},
+                  h("svg", {className: "slds-icon", viewBox: "0 0 24 24"},
+                    h("path", {
+                      d: `
+                      M11 9c-.5 0-1-.5-1-1s.5-1 1-1 1 .5 1 1-.5 1-1 1z
+                      m1 5.8c0 .2-.1.3-.3.3h-1.4c-.2 0-.3-.1-.3-.3v-4.6c0-.2.1-.3.3-.3h1.4c.2.0.3.1.3.3z
+                      M11 3.8c-4 0-7.2 3.2-7.2 7.2s3.2 7.2 7.2 7.2s7.2-3.2 7.2-7.2s-3.2-7.2-7.2-7.2z
+                      m0 12.5c-2.9 0-5.3-2.4-5.3-5.3s2.4-5.3 5.3-5.3s5.3 2.4 5.3 5.3-2.4 5.3-5.3 5.3z
+                      M 17.6 15.9c-.2-.2-.3-.2-.5 0l-1.4 1.4c-.2.2-.2.3 0 .5l4 4c.2.2.3.2.5 0l1.4-1.4c.2-.2.2-.3 0-.5z
+                      `})
+                  )
+                )
               )
             ),
             h("div", {className: "conf-line radio-buttons"},
@@ -1205,7 +1238,7 @@ class App extends React.Component {
             h("div", {className: "conf-line"},
               h("label", {className: "conf-input", title: "The number of records per batch. A higher value is faster but increases the risk of errors due to governor limits."},
                 h("span", {className: "conf-label"}, "Batch size"),
-                h("span", {className: "conf-value button-space"},
+                h("span", {className: "conf-value"},
                   h("input", {type: "number", value: model.batchSize, onChange: this.onBatchSizeChange, className: (model.batchSizeError() ? "confError" : "") + " batch-size"}),
                   h("div", {className: "conf-error", hidden: !model.batchSizeError()}, model.batchSizeError())
                 )
@@ -1213,7 +1246,9 @@ class App extends React.Component {
               h("label", {className: "conf-input", title: "The number of batches to execute concurrently. A higher number is faster but increases the risk of errors due to lock congestion."},
                 h("span", {className: "conf-label"}, "Threads"),
                 h("span", {className: "conf-value"},
-                  h("input", {type: "number", value: model.batchConcurrency, onChange: this.onBatchConcurrencyChange, className: (model.batchConcurrencyError() ? "confError" : "") + " batch-size"}),
+                  h("button", {className: "change-value decreaser", title: "Decrease the value by 1", onClick: this.clickDecrease, "data-targetid": "batchConcurrency"}, "-"),
+                  h("input", {id: "batchConcurrency", type: "text", value: model.batchConcurrency, onChange: this.onBatchConcurrencyChange, className: (model.batchConcurrencyError() ? "confError" : "") + " batch-size"}),
+                  h("button", {className: "change-value increaser", title: "Increase the value by 1", onClick: this.clickIncrease, "data-targetid": "batchConcurrency"}, "+"),
                   h("span", {hidden: !model.isWorking()}, model.activeBatches),
                   h("div", {className: "conf-error", hidden: !model.batchConcurrencyError()}, model.batchConcurrencyError())
                 )