+++ CHANGES.md
@@ -17,6 +17,7 @@
 - Add option to colorize sandbox banner in the same color as the favicon [doc](https://tprouvot.github.io/Salesforce-Inspector-reloaded/how-to/?h=favico#customize-sandbox-banner-color)
 - Add "Object Access" link in Object popup tab to display Object permission details (Winter 25 feature)
 - Fix Platform Event links on popup [issue 500](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/500)
+- New Dark Mode with Accent colors available from Option page [feature 64](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/64) (contribution by [AstIsMe](https://github.com/Astisme))
 
 ## Version 1.24